% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expm.r
\name{expm}
\alias{expm}
\title{Matrix Exponentiation}
\usage{
expm(x, p = 6)
}
\arguments{
\item{x}{A numeric matrix.}

\item{p}{Order of the Pade' approximation.  Should be between 1 and 13.
6 is typical.}
}
\value{
exp(x)
}
\description{
Matrix exponentiation via Pade' approximations.
}

