% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libexpm-package.r
\docType{package}
\name{libexpm-package}
\alias{libexpm-package}
\title{Matrix Exponentiation}
\description{
\tabular{ll}{
   Package: \tab memuse \cr
   Type: \tab Package \cr
   License: \tab BSD 2-Clause \cr
   NeedsCompilation: \tab yes
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{http://github.com/wrathematics/libexpm}
}
\keyword{Package}

