% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normest.r
\name{normest}
\alias{normest}
\title{Norm Estimation}
\usage{
normest(x, pow)
}
\arguments{
\item{x}{A numeric matrix.}

\item{pow}{An integer power.}
}
\value{
exp(x)
}
\description{
Estimates the 1-norm of a power of a matrix.
}

